+++ src/main/java/com/capstone/HisTour/HisTourApplication.java
@@ -2,10 +2,12 @@ package com.capstone.HisTour;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.cache.annotation.EnableCaching;
 import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
 
 @SpringBootApplication
 @EnableJpaAuditing
+@EnableCaching
 public class HisTourApplication {
 
 	public static void main(String[] args) {
